finish coding and begin to modify to ci/cd pipeline section

+++ springboot-react-fullstack/src/main/java/com/cloud/fullstack/student/StudentService.java
@@ -1,5 +1,7 @@
 package com.cloud.fullstack.student;
 
+import com.cloud.fullstack.exception.BadRequestException;
+import com.cloud.fullstack.exception.StudentNotFoundException;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -15,11 +17,17 @@ public class StudentService {
     }
 
     public void addStudent(Student student) {
-        // todo: check if email is taken
+        Boolean isEmailAlreadyTaken = studentRepository.selectExistEmail(student.getEmail());
+        if (isEmailAlreadyTaken) {
+            throw new BadRequestException("Email " + student.getEmail() + " taken");
+        }
         studentRepository.save(student);
     }
 
     public void deleteStudent(String studentId) {
+        if (!studentRepository.existsById(Long.parseLong(studentId))) {
+            throw new StudentNotFoundException("Student " + studentId + " not found");
+        }
         studentRepository.deleteById(Long.parseLong(studentId));
     }
 }
